requirements.txt
pywin32
pillow
tkinterdnd2
